Move layout context to separate comp
