+++ packages/user-interface/src/index.ts
@@ -4,3 +4,4 @@ export * from './contexts';
 export * from './interfaces';
 export * from './hooks';
 export * from './constants';
+export * from './providers';
