+++ packages/user-interface/src/providers/layout-provider/layout-provider.test.tsx
@@ -1,0 +1,7 @@
+import {LayoutProvider} from './layout-provider';
+
+describe('LayoutProvider', () => {
+  it('is truthy', () => {
+    expect(LayoutProvider).toBeTruthy();
+  });
+});
